Add initial garden_bed_hose_guide model
This is one half of the inner part of a torus that will sit over the edge
of a raised garden bed or thin wall and will 'guide' a hose, vine or other
thin thing.  The hose can sit on a nice wide bend and not be pinched.

Two of these printed should then wedge together, with the tang of each
fitting into the opposite mortise.  This means we don't need supports.

Signed-off-by: Paul Wayper <paulway@redhat.com>