Add the square-to-circle polyhedron generator script
This script generates an OpenSCAD polyhedron (post version 2014.3) that
joins a square of internal dimensions `sq_x` and `sq_y` on the X-Y plane to
a circle of internal diameter `circ_diam` at `height` above the plane, in a
shape with `thickness` thick walls, using `divisions` 'slices' of
revolution.  The more slices, the more accurate the polyhedron will be
but the larger the resultant text.

Until such time as I write a module in OpenSCAD that constructs such a
polyhedron on the fly, this is the best way of generating such an
adapter.

Signed-off-by: Paul Wayper <paulway@redhat.com>